added the score calculation part

+++ src/Questionnaire/questionnaire.service.ts
@@ -4,6 +4,7 @@ import { QuestionnaireEntity} from "./questionnaire.entity";
 import { IQuestionnaire,IQuestionnaireService} from "./Interfaces";
 import {SubDomainEntity} from "../DomainForQuestionnaire/SubDomain/subDomain.entity";
 import {DomainEntity} from "../DomainForQuestionnaire/Domain/domain.entity";
+import {async} from "rxjs/scheduler/async";
 
 @Component()
 export class QuestionnaireService implements IQuestionnaireService{
@@ -39,7 +40,6 @@ export class QuestionnaireService implements IQuestionnaireService{
          await getConnection().createQueryBuilder().relation(QuestionnaireEntity,"subdomain")
              .of(selectedQuestionnaire.id).set(questionnaire.subdomain);
          return await this.questionnaireRepository.findOneById(selectedQuestionnaire.id);
-
     }
     public async updateQuestionnaire(id:number,newQuestionnaire:IQuestionnaire):Promise<QuestionnaireEntity|null>{
         const questionnaire = await this.questionnaireRepository.findOneById(id);
@@ -50,12 +50,19 @@ export class QuestionnaireService implements IQuestionnaireService{
         return await this.questionnaireRepository.findOneById(id);
     }
     public async deleteQuestionnaire(id:number):Promise<string>{
-        await this.questionnaireRepository.deleteById(id);
-        const questionnaire = this.questionnaireRepository.findOneById(id);
-        if(questionnaire){
-            return 'delete fail';
-        }else{
+        // await this.questionnaireRepository.deleteById(id);
+        // const questionnaire = this.questionnaireRepository.findOneById(id);
+        // if(questionnaire){
+        //     return 'delete fail';
+        // }else{
+        //     return 'delete success';
+        // }
+        const seletedQuestionnaire = await this.questionnaireRepository.findOneById(id);
+        if(seletedQuestionnaire){
+            await this.questionnaireRepository.deleteById(id);
             return 'delete success';
+        }else{
+            return 'delete fail';
         }
     }
 
@@ -76,5 +83,37 @@ export class QuestionnaireService implements IQuestionnaireService{
             await resultSubDomains.push(subDomainWithQ);
         }
         return await {selectedDomain,resultSubDomains};
+    }
+
+    public async calculateDomainMaxAndMin(){
+        const selectedDomains = await getConnection().createQueryBuilder().select().from(DomainEntity,"domain").getMany();
+        let domains = await selectedDomains.map((domain)=>{
+            return {domain:domain.domain};
+        });
+        const questionnaires = await getRepository(QuestionnaireEntity).createQueryBuilder("questionnaire")
+            .leftJoinAndSelect("questionnaire.domain","domain")
+            .getMany();
+        await domains.forEach(async(domainItem)=>{
+            let maxScore:number = 0;
+            let minScore:number = 0;
+            let questionnairesGroupByDomain = await questionnaires.filter((q)=>q.domain.domain === domainItem.domain);
+            await questionnairesGroupByDomain.forEach(async(questionnaire)=>{
+                let maxPoint:number;
+                let minPoint:number;
+                let array:number[] = [];
+                questionnaire.options.forEach((option)=>{
+                    array.push(option.point);
+                })
+                maxPoint = Math.max(...array);
+                maxScore = maxPoint * questionnaire.weight;
+                minPoint = Math.min(...array);
+                minScore = minPoint * questionnaire.weight;
+            });
+            await getConnection().createQueryBuilder().update(DomainEntity)
+                .set({maxScore:maxScore}).where("domain = :domain",{domain:domainItem.domain}).execute();
+
+            await getConnection().createQueryBuilder().update(DomainEntity)
+                .set({minScore:minScore}).where("domain = :domain",{domain:domainItem.domain}).execute();
+        });
     }
 }
